Fix capability reporting instability
Previously if we immediately tried to set pin mode, we would crash
because the pins vector would be empty.
The fix for this was to wait for the capability response before we allow
the board constructor to return.
In addition to this we also wait for the analog mapping response and the
firmware response because why not.